requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bidict==0.22.1
click==8.1.7
fastapi==0.109.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
pydantic==2.6.0
pydantic_core==2.16.1
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
wsproto==1.2.0
